Create README.md
Here is a basic outline of how you could structure your application:

User Interface (UI): The app should have a user-friendly interface that allows users to take or upload photos of lost dogs. There should also be options to add filters such as dog size, color, and type.

Geolocation: The application must be able to obtain the user's current location. This can be achieved using the device's GPS.

Image processing: Once the photo is taken, the application must be able to process it. This could involve using artificial intelligence to identify specific characteristics of the dog in the image.

Database: The application must have a database to store the information of lost dogs. Each entry in the database could contain the image of the dog, the location where it was found, and any other relevant information.

Notifications: If a possible match is found, the application should notify the user.

Open Source: Since you mentioned that you want the app to be open source, you can host your code on platforms like GitHub so that other developers can contribute to the project.